@@ -26,7 +26,7 @@
 		"prettier-plugin-svelte": "^3.2.5",
 		"prettier": "^3.3.2",
 		"svelte-check": "^3.8.2",
-		"svelte": "^4.2.18",
+		"svelte": "^5.0.0-next.165",
 		"tslib": "^2.6.3",
 		"typescript-eslint": "^8.0.0-alpha.20",
 		"typescript": "^5.5.2",
