@@ -35,7 +35,7 @@
 		"prettier": "^3.3.2",
 		"prettier-plugin-svelte": "^3.2.5",
 		"prettier-plugin-tailwindcss": "^0.6.5",
-		"svelte": "^4.2.12",
+		"svelte": "^5.0.0-next.165",
 		"svelte-check": "^3.8.1",
 		"tailwind-merge": "^2.3.0",
 		"tailwindcss": "^3.4.4",
