@@ -1,56 +1,56 @@
-<script>
-	import welcome_fallback from "$lib/images/svelte-welcome.png";
-	import welcome from "$lib/images/svelte-welcome.webp";
-</script>
-
 <svelte:head>
-	<title>Home</title>
-	<meta name="description" content="Svelte demo app" />
+	<title>SVA</title>
+	<meta name="description" content="Svelte App Starter" />
 </svelte:head>
 
 <section>
 	<h1>
-		<span class="welcome">
-			<picture>
-				<source srcset={welcome} type="image/webp" />
-				<img src={welcome_fallback} alt="Welcome" />
-			</picture>
-		</span>
-
-		to your new<br />SvelteKit app
+		<span>Great Work.</span>
+		<br />
+		<mark>You're all set!</mark>
 	</h1>
 
-	<h2>
-		try editing <strong>src/routes/+page.svelte</strong>
-	</h2>
+	<h2>Useful Documentation</h2>
+
+	<ul>
+		<li><a target="_blank" href="https://svelte.dev/">Svelte</a></li>
+		<li><a target="_blank" href="https://kit.svelte.dev/">SvelteKit</a></li>
+		<li><a target="_blank" href="https://www.typescriptlang.org/">Typescript</a></li>
+		<li><a target="_blank" href="https://tailwindcss.com">Tailwind CSS</a></li>
+		<li><a target="_blank" href="https://lucia-auth.com/">Lucia</a></li>
+		<li><a target="_blank" href="https://orm.drizzle.team/">Drizzle</a></li>
+		<li><a target="_blank" href="https://typicode.github.io/husky/">Husky</a></li>
+		<li><a target="_blank" href="https://prettier.io/">Prettier</a></li>
+		<li><a target="_blank" href="https://eslint.org/">ESLint</a></li>
+	</ul>
 </section>
 
 <style>
-	section {
-		display: flex;
-		flex-direction: column;
-		justify-content: center;
-		align-items: center;
-		flex: 0.6;
-	}
-
 	h1 {
 		width: 100%;
+		color: white;
+		font-size: 3rem;
+		line-height: 1.05;
+		font-weight: 600;
+		margin: 1rem 0;
 	}
 
-	.welcome {
-		display: block;
-		position: relative;
-		width: 100%;
-		height: 0;
-		padding: 0 0 calc(100% * 495 / 2048) 0;
+	h1 mark {
+		background: none;
+		color: rgb(0, 255, 0);
 	}
 
-	.welcome img {
-		position: absolute;
-		width: 100%;
-		height: 100%;
-		top: 0;
-		display: block;
+	h2 {
+		color: rgba(255, 255, 255, 0.6);
+		margin: 1rem 0;
+	}
+
+	ul {
+		font-size: 1.5rem;
+		line-height: 1.5;
+	}
+	ul a:hover {
+		color: rgb(0, 255, 0);
+		text-decoration: underline;
 	}
 </style>
