@@ -1,4 +1,5 @@
 import { dev } from '$app/environment';
+import { DATABASE_URL } from '$env/static/private';
 import * as schema from '$lib/server/schema';
 import Database from 'better-sqlite3';
 import { drizzle } from 'drizzle-orm/better-sqlite3';
@@ -13,7 +14,7 @@ const global_db = globalThis as unknown as {
 
 // We're using better-sqlite3 to support a local sqlite database.
 // You might want to switch to the libsql adapter for production.
-const connection = global_db.connection ?? new Database('sqlite.db');
+const connection = global_db.connection ?? new Database(DATABASE_URL);
 if (dev) global_db.connection = connection;
 
 // You can find specific SQLite adapters in the Drizzle docs.
