@@ -66,7 +66,7 @@ export async function GET(event: RequestEvent): Promise<Response> {
 		}
 		return new Response(null, {
 			status: 302,
-			headers: { Location: '/' },
+			headers: { Location: '/account' },
 		});
 	} catch (e) {
 		console.warn('Some auth error: ', e);
