@@ -1,20 +1,46 @@
+<svelte:head>
+	<title>SVA</title>
+	<meta name="description" content="Svelte App Starter" />
+</svelte:head>
+
 <section>
-	<a href="/login/github"> Login with GitHub </a>
+	<h1>
+		Lucia <mark>Login</mark>
+	</h1>
+
+	<h2>Providers</h2>
+
+	<ul>
+		<li><a href="/login/github"> Login with GitHub </a></li>
+	</ul>
 </section>
 
-<style lang="postcss">
-	section {
+<style>
+	h1 {
 		width: 100%;
-		min-height: 100vh;
-		display: grid;
-		place-items: center;
+		color: white;
+		font-size: 3rem;
+		line-height: 1.05;
+		font-weight: 600;
+		margin: 1rem 0;
 	}
 
-	a {
-		color: blue;
+	h1 mark {
+		background: none;
+		color: rgb(0, 255, 0);
 	}
 
-	a:hover {
+	h2 {
+		color: rgba(255, 255, 255, 0.6);
+		margin: 1rem 0;
+	}
+
+	ul {
+		font-size: 1.5rem;
+		line-height: 1.5;
+	}
+	ul a:hover {
+		color: rgb(0, 255, 0);
 		text-decoration: underline;
 	}
 </style>
